refactor(Schedule): move WeekNavigator into Schedule
